tests: Move staging index update coverage
The staging index update helper now lives in staging.index_update. Tests
for writing generated buffers into the Git index should sit beside that
module rather than inside the buffer construction tests.

The old staging operations test mixed content-buffer cases with repository
index mutation cases. After the source extraction, that test still imported
the moved helper from staging.operations.

This commit addresses that by moving the index update fixture and tests
into tests/staging/test_index_update.py and leaving
tests/staging/test_operations.py focused on target buffer construction.

Architecture guard updates will follow in a separate tests commit.
